refactor: переименование логотипов проекта
